+++ src/spin_pulse/environment/noise/white.py
@@ -37,7 +37,7 @@ class WhiteNoiseTimeTrace(NoiseTimeTrace):
     """
 
     def __init__(
-        self, T2S: int, duration: int, segment_duration: int, seed: int | None = None
+        self, T2S: float, duration: int, segment_duration: int, seed: int | None = None
     ):
         r"""Initialize a white noise time trace for spin qubit simulations.
